Refactor: ViewModel 분리
